# patches/controller-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSC SRE Automacao Controller",
    "description": "API do Controller de automacao SRE. Orquestra a execucao de automacoes nos Agents distribuidos por cluster, gerencia logs de execucao, autenticacao JWT e integracao com o OAS (Orquestrador de Automacao de Servicos).",
    "version": "3.9.6",
    "contact": {
      "name": "PSC SRE Automacao"
    }
  },
  "servers": [
    {
      "url": "https://sre-aut-controller.psc.{cluster}.bb.com.br",
      "description": "Controller por cluster",
      "variables": {
        "cluster": {
          "description": "Nome do cluster OpenShift",
          "default": "k8shmlbb111b"
        }
      }
    }
  ],
...
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Emitir token JWT",
        "description": "Emite um token JWT para autenticacao nas rotas protegidas. Requer API Key no header x-api-key.",
        "operationId": "issueToken",
        "security": [{ "apiKeyAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": { "type": "string", "example": "helmfire-ritmo-client" },
                  "scope": { "oneOf": [{ "type": "string", "example": "execute:automation send:logs" }, { "type": "array", "items": { "type": "string" } }] },
                  "expiresIn": { "type": "string", "example": "5m" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token emitido", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } } },
          "400": { "description": "Scopes invalidos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "API Key invalida", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Scopes nao permitidos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Token JWT emitido por POST /auth/token. Algoritmo HS256, issuer=psc-sre-automacao-controller." },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "API Key para emissao de tokens JWT e consulta de scopes." },
      "internalOrigin": { "type": "apiKey", "in": "header", "name": "x-techbb-namespace", "description": "Autenticacao por origem interna do Kubernetes." }
    },
    "schemas": {
      "ErrorResponse": { "type": "object", "properties": { "ok": { "type": "boolean", "example": false }, "error": { "type": "string" }, "detail": { "type": "string" }, "execId": { "type": "string" } } },
...
      "TokenResponse": { "type": "object", "properties": { "token": { "type": "string" }, "tokenType": { "type": "string", "example": "Bearer" }, "expiresIn": { "type": "string" }, "howToUse": { "type": "string" } } },
...
    }
  }
}